requirements-dev.txt
-r requirements.txt
pytest>=8.0
pytest-django>=4.8
playwright>=1.50
